requirements.txt
python-binance
python-telegram-bot==20.6
pandas
ta
